move netcdf handling functions out of utilities
